Move `SegmentedMessage` to its own file
